requirements.txt
asn1crypto==1.5.1
bcrypt==4.1.2
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
filelock==3.13.1
Flask==3.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pip==23.3.2
platformdirs==3.11.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
pytz==2023.3.post1
requests==2.31.0
setuptools==69.0.3
snowflake-connector-python==3.6.0
sortedcontainers==2.4.0
tomlkit==0.12.3
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
wheel==0.42.0